% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignmentMatrix.R
\name{alignmentMatrix}
\alias{alignmentMatrix}
\title{Aligned topic distribution across different chains}
\usage{
alignmentMatrix(
  theta,
  sce,
  K,
  iter = 2000,
  chain = 4,
  SampleID_name = "cell_id"
)
}
\arguments{
\item{theta}{Array. Stan output results of topic proportion.}

\item{sce}{SingleCellExperiment.}

\item{K}{Integer. Number of topics}

\item{iter}{Integer. Iteration used for each chain, including warm-up.}

\item{chain}{Integer. Number of chains used.}

\item{SampleID_name}{Character. The variable indicate sample ID.}
}
\value{
Matrix.
}
\description{
Aligned topic distribution across different chains
}
